+++ binsync/__init__.py
@@ -1,12 +1,15 @@
+VERSION = "2.3.0"
+
 import logging
+
 logging.getLogger("binsync").addHandler(logging.NullHandler())
-from .loggercfg import Loggers
+from binsync.loggercfg import Loggers
+
 loggers = Loggers()
 del Loggers
 del logging
 
-from .state import State, ArtifactType
-from .client import Client, StateContext, ConnectionWarnings
-from .data import *
-
-VERSION = "2.3.0"
+from binsync import common
+from binsync.client import Client, ConnectionWarnings, StateContext
+from binsync.data import *
+from binsync.state import ArtifactType, State
