+++ binsync/common/ui/__init__.py
@@ -1,15 +1,3 @@
-ui_version = "PySide2"
-
-
-def set_ui_version(version):
-    global ui_version
-    valid_version = [
-        "PyQt5",
-        "PySide2",
-        "PySide6"
-    ]
-
-    if version in valid_version:
-        ui_version = version
-    else:
-        raise Exception("Failed to set BinSync UI version")
+from binsync.common.ui import config_dialog
+from binsync.common.ui import control_panel
+from binsync.common.ui import panel_tabs
