+++ binsync/common/ui/qt_objects.py
@@ -1,0 +1,74 @@
+from binsync.common.ui.version import ui_version
+
+if ui_version == "PySide2":
+    from PySide2.QtCore import QDir, Qt, Signal
+    from PySide2.QtWidgets import (
+        QAbstractItemView,
+        QCheckBox,
+        QComboBox,
+        QDialog,
+        QFileDialog,
+        QGridLayout,
+        QGroupBox,
+        QHBoxLayout,
+        QHeaderView,
+        QLabel,
+        QLineEdit,
+        QMenu,
+        QMessageBox,
+        QPushButton,
+        QStatusBar,
+        QTableWidget,
+        QTableWidgetItem,
+        QTabWidget,
+        QVBoxLayout,
+        QWidget,
+    )
+elif ui_version == "PySide6":
+    from PySide6.QtCore import QDir, Qt, Signal
+    from PySide6.QtWidgets import (
+        QAbstractItemView,
+        QCheckBox,
+        QComboBox,
+        QDialog,
+        QFileDialog,
+        QGridLayout,
+        QGroupBox,
+        QHBoxLayout,
+        QHeaderView,
+        QLabel,
+        QLineEdit,
+        QMenu,
+        QMessageBox,
+        QPushButton,
+        QStatusBar,
+        QTableWidget,
+        QTableWidgetItem,
+        QTabWidget,
+        QVBoxLayout,
+        QWidget,
+    )
+else:
+    from PyQt5.QtCore import QDir, Qt, Signal
+    from PyQt5.QtWidgets import (
+        QAbstractItemView,
+        QCheckBox,
+        QComboBox,
+        QDialog,
+        QFileDialog,
+        QGridLayout,
+        QGroupBox,
+        QHBoxLayout,
+        QHeaderView,
+        QLabel,
+        QLineEdit,
+        QMenu,
+        QMessageBox,
+        QPushButton,
+        QStatusBar,
+        QTableWidget,
+        QTableWidgetItem,
+        QTabWidget,
+        QVBoxLayout,
+        QWidget,
+    )
