+++ binsync/common/ui/utils.py
@@ -1,16 +1,7 @@
-from . import ui_version
-import logging
 import datetime
+import logging
 
-if ui_version == "PySide2":
-    from PySide2.QtWidgets import QTableWidgetItem
-    from PySide2.QtCore import Qt
-elif ui_version == "PySide6":
-    from PySide6.QtWidgets import QTableWidgetItem
-    from PySide6.QtCore import Qt
-else:
-    from PyQt5.QtWidgets import QTableWidgetItem
-    from PyQt5.QtCore import Qt
+from binsync.common.ui.qt_objects import Qt, QTableWidgetItem
 
 l = logging.getLogger(__name__)
