+++ binsync/data/__init__.py
@@ -1,9 +1,9 @@
-from .user import User
-from .artifact import Artifact
-from .func import Function, FunctionHeader, FunctionArgument
-from .comment import Comment
-from .patch import Patch
-from .stack_variable import StackVariable, StackOffsetType
-from .struct import StructMember, Struct
-from .global_variable import GlobalVariable
-from .enum import Enum
+from binsync.data.user import User
+from binsync.data.artifact import Artifact
+from binsync.data.func import Function, FunctionHeader, FunctionArgument
+from binsync.data.comment import Comment
+from binsync.data.patch import Patch
+from binsync.data.stack_variable import StackVariable, StackOffsetType
+from binsync.data.struct import StructMember, Struct
+from binsync.data.global_variable import GlobalVariable
+from binsync.data.enum import Enum
