+++ binsync/data/global_variable.py
@@ -1,5 +1,6 @@
-from .artifact import Artifact
 import toml
+
+from binsync.data.artifact import Artifact
 
 
 class GlobalVariable(Artifact):
