+++ binsync/data/patch.py
@@ -1,7 +1,8 @@
 import codecs
+
 import toml
 
-from .artifact import Artifact
+from binsync.data.artifact import Artifact
 
 
 class Patch(Artifact):
