+++ binsync/data/stack_variable.py
@@ -1,6 +1,6 @@
 import toml
 
-from .artifact import Artifact
+from binsync.data.artifact import Artifact
 
 
 class StackOffsetType:
