+++ plugins/angr_binsync/__init__.py
@@ -1,1 +1,1 @@
-from .binsync_plugin import BinSyncPlugin
+from angrmanagement.plugins.angr_binsync.binsync_plugin import BinSyncPlugin
