+++ plugins/angr_binsync/control_panel_view.py
@@ -1,14 +1,9 @@
 import logging
-from PySide2.QtWidgets import QVBoxLayout
 
-from binsync.common.ui import set_ui_version
-
-from .controller import AngrBinSyncController
-from ...ui.views.view import BaseView
-
-set_ui_version("PySide2")
-# pylint: disable=wrong-import-position,wrong-import-order
+from angrmanagement.plugins.angr_binsync.controller import AngrBinSyncController
+from angrmanagement.ui.views.view import BaseView
 from binsync.common.ui.control_panel import ControlPanel
+from binsync.common.ui.qt_objects import QVBoxLayout
 
 
 l = logging.getLogger(__name__)
