+++ plugins/binja_binsync/binja_binsync.py
@@ -20,7 +20,7 @@ from collections import defaultdict
 import logging
 
 import binsync
-from binsync.common.ui import set_ui_version
+from binsync.common.ui.version import set_ui_version
 set_ui_version("PySide6")
 from binsync.common.ui.config_dialog import SyncConfig
 from binsync.common.ui.control_panel import ControlPanel
